fix(ui5-library-inquirer): code review
